+++ src/Vortice.Direct3D12/ID3D12Device.cs
@@ -246,7 +246,7 @@ namespace Vortice.Direct3D12
             return new ID3D12GraphicsCommandList(nativePtr);
         }
 
-        public ID3D12Fence CreateFence(long initialValue, FenceFlags flags = FenceFlags.None)
+        public ID3D12Fence CreateFence(ulong initialValue, FenceFlags flags = FenceFlags.None)
         {
             return CreateFence(initialValue, flags, typeof(ID3D12Fence).GUID);
         }
@@ -407,7 +407,7 @@ namespace Vortice.Direct3D12
 
         public ID3D12Resource CreatePlacedResource(
             ID3D12Heap heap,
-            long heapOffset,
+            ulong heapOffset,
             ResourceDescription resourceDescription,
             ResourceStates initialState,
             ClearValue? clearValue = null)