+++ src/Vortice.Direct3D12/ResourceDescription.cs
@@ -23,11 +23,11 @@ namespace Vortice.Direct3D12
         /// <param name="optionFlags"></param>
         public ResourceDescription(
             ResourceDimension dimension,
-            long alignment,
-            long width,
+            ulong alignment,
+            ulong width,
             int height,
-            short depthOrArraySize,
-            short mipLevels,
+            ushort depthOrArraySize,
+            ushort mipLevels,
             Format format,
             int sampleCount,
             int sampleQuality,
@@ -55,44 +55,47 @@ namespace Vortice.Direct3D12
                 1, 1, 1, Format.Unknown, 1, 0, TextureLayout.RowMajor, flags);
         }
 
-        public static ResourceDescription Buffer(long width, ResourceFlags flags = ResourceFlags.None, long alignment = 0)
+        public static ResourceDescription Buffer(
+            ulong width, 
+            ResourceFlags flags = ResourceFlags.None, 
+            ulong alignment = 0)
         {
             return new ResourceDescription(ResourceDimension.Buffer, alignment, width, 1, 1, 1, Format.Unknown, 1, 0, TextureLayout.RowMajor, flags);
         }
 
         public static ResourceDescription Texture1D(Format format,
-            long width,
-            short arraySize = 1,
-            short mipLevels = 0,
+            ulong width,
+            ushort arraySize = 1,
+            ushort mipLevels = 0,
             ResourceFlags flags = ResourceFlags.None,
             TextureLayout layout = TextureLayout.Unknown,
-            long alignment = 0)
+            ulong alignment = 0)
         {
             return new ResourceDescription(ResourceDimension.Texture1D, alignment, width, 1, arraySize, mipLevels, format, 1, 0, layout, flags);
         }
 
         public static ResourceDescription Texture2D(Format format,
-            long width,
+            ulong width,
             int height,
-            short arraySize = 1,
-            short mipLevels = 0,
+            ushort arraySize = 1,
+            ushort mipLevels = 0,
             int sampleCount = 1,
             int sampleQuality = 0,
             ResourceFlags flags = ResourceFlags.None,
             TextureLayout layout = TextureLayout.Unknown,
-            long alignment = 0)
+            ulong alignment = 0)
         {
             return new ResourceDescription(ResourceDimension.Texture2D, alignment, width, height, arraySize, mipLevels, format, sampleCount, sampleQuality, layout, flags);
         }
 
         public static ResourceDescription Texture3D(Format format,
-            long width,
+            ulong width,
             int height,
-            short depth,
-            short mipLevels = 0,
+            ushort depth,
+            ushort mipLevels = 0,
             ResourceFlags flags = ResourceFlags.None,
             TextureLayout layout = TextureLayout.Unknown,
-            long alignment = 0)
+            ulong alignment = 0)
         {
             return new ResourceDescription(ResourceDimension.Texture3D, alignment, width, height, depth, mipLevels, format, 1, 0, layout, flags);
         }